реализовал миграцию на реальную базу данных (MS SQL Server) пофиксил логику с удалением аватара

+++ RazorPagesGeneral/Pages/Employees/Edit.cshtml.cs
@@ -1,8 +1,5 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -56,8 +53,11 @@ namespace RazorPagesGeneral.Pages.Employees
                     if(Employee.PhotoPath != null)
                     {
                         string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Employee.PhotoPath);
-                        System.IO.File.Delete(filePath);
 
+                        if (Employee.PhotoPath != "noimage.png")
+                        {
+                            System.IO.File.Delete(filePath);
+                        }
                     }
 
                     Employee.PhotoPath = ProcessUploadFile();
